28: Improved loading time  from 4s to 12 ms
This commit contains changes to pictures that slowed down the website considerably.  Accomplished with CSS sprites and resizing gigantic png's into webp. The webp formats have 2 sizes that will be pulled based on the density of the screen. Also here is the deletion of sandbox.html from master, as it's used as a testing area.